+++ src/document/toHtml.js
@@ -1,5 +1,5 @@
 
-//
+//turn a Doc object into a HTML string
 const toHtml = function(doc, options) {
   let data = doc.data;
   let html = '';
